# Conflicts:
#	WebAPI/example.log
#	WebAPI/example.log.2024-05-20.1

+++ Application/GroupPermissions/CommandHandlers/DeleteGroupPermissionCommandHandler.cs
@@ -2,15 +2,19 @@
 using Application.GroupPermissions.Dto;
 using AutoMapper;
 using Common.Exceptions;
-using Domain.Entities;
 using Infrastructure;
 using MediatR;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
 
 namespace Application.GroupPermissions.CommandHandlers
 {
     public class DeleteGroupPermissionCommandHandler : IRequestHandler<DeleteGroupPermissionCommand, GroupPermissionDto>
     {
-        private BanHangContext _context;
+        private readonly BanHangContext _context;
         private readonly IMapper _mapper;
 
         public DeleteGroupPermissionCommandHandler(BanHangContext context, IMapper mapper)
@@ -18,14 +22,18 @@ namespace Application.GroupPermissions.CommandHandlers
             _context = context;
             _mapper = mapper;
         }
+
         public async Task<GroupPermissionDto> Handle(DeleteGroupPermissionCommand request, CancellationToken cancellationToken)
         {
-            GroupPermission? groupPermission = _context.GroupPermissions.Find(request.Id, cancellationToken) ?? throw new AppException(
-                ExceptionCode.Notfound,
-                "Không tìm thấy GroupPermission "
-                );
+            var groupPermission = await _context.GroupPermissions.FindAsync(request.Id);
+            if (groupPermission == null)
+            {
+                throw new AppException(ExceptionCode.Notfound, "Không tìm thấy GroupPermission");
+            }
+
             _context.GroupPermissions.Remove(groupPermission);
-            _context.SaveChanges();
+            await _context.SaveChangesAsync(cancellationToken);
+
             return _mapper.Map<GroupPermissionDto>(groupPermission);
         }
     }
